common.postinst: continue on error
In the postinst script, if we hit an error building for one kernel,
record the error and continue on to the next kernel. When done, exit
with the value of any recorded error. This allows us to do a best-
effort build for all installed kernels. Allow this feature to be
turned off if this behavior isn't desired.